+++ flink-core/src/test/java/org/apache/flink/testutils/TestFileSystem.java
@@ -19,9 +19,7 @@
 package org.apache.flink.testutils;
 
 import java.io.IOException;
-import java.lang.reflect.Field;
 import java.net.URI;
-import java.util.Map;
 
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.core.fs.FSDataInputStream;
@@ -32,8 +30,14 @@ import org.apache.flink.core.fs.Path;
 import org.apache.flink.core.fs.local.LocalFileStatus;
 import org.apache.flink.core.fs.local.LocalFileSystem;
 
+/**
+ * A test file system. This also has a service entry in the test
+ * resources, to be loaded during tests.
+ */
 public class TestFileSystem extends LocalFileSystem {
-	
+
+	public static final String SCHEME = "test";
+
 	private static int streamOpenCounter;
 	
 	public static int getNumtimeStreamOpened() {
@@ -74,24 +78,17 @@ public class TestFileSystem extends LocalFileSystem {
 
 	@Override
 	public URI getUri() {
-		return URI.create("test:///");
-	}
-
-	public static void registerTestFileSysten() throws Exception {
-		Class<FileSystem> fsClass = FileSystem.class;
-		Field dirField = fsClass.getDeclaredField("FS_FACTORIES");
-
-		dirField.setAccessible(true);
-		@SuppressWarnings("unchecked")
-		Map<String, FileSystemFactory> map = (Map<String, FileSystemFactory>) dirField.get(null);
-		dirField.setAccessible(false);
-
-		map.put("test", new TestFileSystemFactory());
+		return URI.create(SCHEME + ":///");
 	}
 
 	// ------------------------------------------------------------------------
 
-	private static final class TestFileSystemFactory implements FileSystemFactory {
+	public static final class TestFileSystemFactory implements FileSystemFactory {
+
+		@Override
+		public String getScheme() {
+			return SCHEME;
+		}
 
 		@Override
 		public void configure(Configuration config) {}
