+++ tools/travis_mvn_watchdog.sh
@@ -301,6 +301,23 @@ check_shaded_artifacts() {
 		return 1
 	fi
 
+
+	HADOOP=`cat allClasses | grep '^org/apache/hadoop' | wc -l`
+	if [ "$HADOOP" != "0" ]; then
+		echo "=============================================================================="
+		echo "Detected '$HADOOP' Hadoop classes in the dist jar
+		echo "=============================================================================="
+		return 1
+	fi
+
+	MAPR=`cat allClasses | grep '^com/mapr' | wc -l`
+	if [ "$MAPR" != "0" ]; then
+		echo "=============================================================================="
+		echo "Detected '$MAPR' MapR classes in the dist jar
+		echo "=============================================================================="
+		return 1
+	fi
+
 	return 0
 }
